requirements.txt
dash
flask_login
flask_sqlalchemy
flask_bcrypt
Flask-WTF
pymssql
pyodbc
email-validator
SQLAlchemy
pandas 
itsdangerous
pyarrow
plotly
dash-bootstrap-components
dash-bootstrap-templates
pytz
requests
python-dotenv
psycopg2
apscheduler